+++ cmd/report.go
@@ -16,9 +16,13 @@ type ReportCmd struct {
 	End            time.Time `short:"t" optional:"" format:"2006-01-02" help:"Date to end report from (default: start + 1d)"`
 	DurationFormat string    `enum:"default,hour" default:"default" help:"How to format duration in output (values: default,hour)" env:"DURATION_FORMAT"`
 	// TODO: Make this a cleaner API
-	Output          report.OutputFormat `short:"o" enum:"default,markdown,html" default:"default" help:"Premade formats (default,markdown,html)"`
-	Template        string              `type:"existingfile" help:"File text/template to use in making report (Overrides --output)"`
-	templateContent string              `kong:"-"`
+	Output   report.OutputFormat `short:"o" enum:"default,markdown,html" default:"default" help:"Premade formats (default,markdown,html)"`
+	Template string              `type:"existingfile" help:"File text/template to use in making report (Overrides --output)"`
+	// TODO: This gets tricky since it could also be time...
+	// TODO: Also probably should be an ENV?
+	Until           time.Duration         `short:"u" help:"Desired duration of work to log"`
+	templateContent string                `kong:"-"`
+	reportOptions   []report.ReportOption `kong:"-"`
 }
 
 func (cmd *ReportCmd) Run(stdout io.Writer, log string) error {
@@ -33,7 +37,7 @@ func (cmd *ReportCmd) Run(stdout io.Writer, log string) error {
 		cmd.templateContent,
 		cmd.Start,
 		filtered,
-		report.WithDurationFormat(cmd.DurationFormat),
+		cmd.reportOptions...,
 	)
 }
 
@@ -61,5 +65,21 @@ func (cmd *ReportCmd) AfterApply() error {
 		}
 		cmd.templateContent = string(bytes)
 	}
+
+	cmd.reportOptions = append(cmd.reportOptions, report.WithDurationFormat(cmd.DurationFormat))
+	if cmd.Until.Seconds() != float64(0) {
+		cmd.reportOptions = append(cmd.reportOptions, report.WithDesiredDurationLogged(cmd.Until))
+	}
+
 	return nil
+}
+
+// TODO: May be shared by other code
+func Today(timestamp time.Time) (time.Time, error) {
+	now := timestamp.Format(time.DateOnly)
+	today, err := time.Parse(time.DateOnly, now)
+	if err != nil {
+		return time.Time{}, err
+	}
+	return today, nil
 }